requirements.txt
arDjango-1> pip freeze
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
sqlparse==0.5.0
tzdata==2024.1